[Mod]: 🔄 Change beta reconstruction reference point to after S4
- Update path length initialization logic for new reference location
- Set time offset reference to TOF S4 layer time
- Change momentum and energy index from 10 to 17 (S4 position)
- Refactor propagation algorithm using loops instead of hardcoded indices
- Fine-tune E=mc² calculations from a new vantage point

+++ include/BetaNL.h
@@ -28,7 +28,7 @@
 class BetaNLPars
 {
 public:
-    static constexpr size_t nTOF = 4;                    // Number of TOF hits
+    static constexpr int nTOF = 4;                       // Number of TOF hits
     static constexpr double SPEED_OF_LIGHT = 29.9792458; // Speed of light in cm/ns
 
 public:
